requirements/common.txt
Flask-SQLAlchemy==1.0

Flask-Script==0.6.3

Flask-Migrate==1.2.0

Flask-Admin==1.0.7

Flask-Login==0.2.9

Flask-WTF==0.9.4

Flask-Gears==0.2

SQLAlchemy==0.8.3

alembic==0.6.2

WTForms==1.0.5

Flask-Babel==0.9

Flask-Babelex==0.9.1

Flask-Mail==0.9.0

python-dateutil==2.2


Gears==0.7.2

gears-clean-css==0.1.2

gears-coffeescript==0.1.3

gears-less==0.3.3

gears-sass==0.2.2

gears-stylus==0.2.1

gears-uglifyjs==0.1.1


celery==3.1.13

flask-thumbnails==0.2.0

Redis